+++ storm-core/test/clj/backtype/storm/messaging/netty_unit_test.clj
@@ -15,7 +15,7 @@
 ;; limitations under the License.
 (ns backtype.storm.messaging.netty-unit-test
   (:use [clojure test])
-  (:import [backtype.storm.messaging TransportFactory])
+  (:import [backtype.storm.messaging TransportFactory IConnection TaskMessage IConnectionCallback])
   (:import [backtype.storm.utils Utils])
   (:use [backtype.storm testing util config log])
   (:use [backtype.storm.daemon.worker :only [is-connection-ready]])
@@ -45,20 +45,37 @@
            (throw (RuntimeException. (str "Netty connections were not ready within " max-wait-ms " ms"))))
          (log-message "All Netty connections are ready"))))))
 
+(defn mk-connection-callback
+  "make an IConnectionCallback"
+  [my-fn]
+  (reify IConnectionCallback
+    (recv [this batch]
+      (doseq [msg batch]
+        (my-fn msg)))))
+
+(defn register-callback
+  "register the local-transfer-fn with the server"
+  [my-fn ^IConnection socket]
+  (.registerRecv socket (mk-connection-callback my-fn)))
+
+(defn- wait-for-not-nil
+  [atm]
+  (while-timeout TEST-TIMEOUT-MS (nil? @atm) (Thread/sleep 10)))
 
 (defn- test-basic-fn [storm-conf]
   (log-message "1. Should send and receive a basic message")
   (let [req_msg (String. "0123456789abcdefghijklmnopqrstuvwxyz")
         context (TransportFactory/makeContext storm-conf)
         port (available-port 6700)
+        resp (atom nil)
         server (.bind context nil port)
+        _ (register-callback (fn [message] (reset! resp message)) server)
         client (.connect context nil "localhost" port)
         _ (wait-until-ready [server client])
-        _ (.send client task (.getBytes req_msg))
-        iter (.recv server 0 0)
-        resp (.next iter)]
-    (is (= task (.task resp)))
-    (is (= req_msg (String. (.message resp))))
+        _ (.send client task (.getBytes req_msg))]
+    (wait-for-not-nil resp)
+    (is (= task (.task @resp)))
+    (is (= req_msg (String. (.message @resp))))
     (.close client)
     (.close server)
     (.term context)))
@@ -88,19 +105,20 @@
   (let [req_msg (String. "0123456789abcdefghijklmnopqrstuvwxyz")
         context (TransportFactory/makeContext storm-conf)
         port (available-port 6700)
+        resp (atom nil)
         server (.bind context nil port)
+        _ (register-callback (fn [message] (reset! resp message)) server)
         client (.connect context nil "localhost" port)
         _ (wait-until-ready [server client])
         _ (.send client task (.getBytes req_msg))
-        iter (.recv server 0 0)
-        resp (.next iter)
         _ (.sendLoadMetrics server {(int 1) 0.0 (int 2) 1.0})
         _ (while-timeout 5000 (empty? (.getLoad client [(int 1) (int 2)])) (Thread/sleep 10))
         load (.getLoad client [(int 1) (int 2)])]
     (is (= 0.0 (.getBoltLoad (.get load (int 1)))))
     (is (= 1.0 (.getBoltLoad (.get load (int 2)))))
-    (is (= task (.task resp)))
-    (is (= req_msg (String. (.message resp))))
+    (wait-for-not-nil resp)
+    (is (= task (.task @resp)))
+    (is (= req_msg (String. (.message @resp))))
     (.close client)
     (.close server)
     (.term context)))
@@ -130,14 +148,15 @@
   (let [req_msg (apply str (repeat 2048000 'c'))
         context (TransportFactory/makeContext storm-conf)
         port (available-port 6700)
+        resp (atom nil)
         server (.bind context nil port)
+        _ (register-callback (fn [message] (reset! resp message)) server)
         client (.connect context nil "localhost" port)
         _ (wait-until-ready [server client])
-        _ (.send client task (.getBytes req_msg))
-        iter (.recv server 0 0)
-        resp (.next iter)]
-    (is (= task (.task resp)))
-    (is (= req_msg (String. (.message resp))))
+        _ (.send client task (.getBytes req_msg))]
+    (wait-for-not-nil resp)
+    (is (= task (.task @resp)))
+    (is (= req_msg (String. (.message @resp))))
     (.close client)
     (.close server)
     (.term context)))
@@ -164,27 +183,28 @@
 
 (defn- test-server-delayed-fn [storm-conf]
   (log-message "4. test server delayed")
-    (let [req_msg (String. "0123456789abcdefghijklmnopqrstuvwxyz")
-          context (TransportFactory/makeContext storm-conf)
-          port (available-port 6700)
-          client (.connect context nil "localhost" port)
+  (let [req_msg (String. "0123456789abcdefghijklmnopqrstuvwxyz")
+        context (TransportFactory/makeContext storm-conf)
+        resp (atom nil)
+        port (available-port 6700)
+        client (.connect context nil "localhost" port)
 
-          server (Thread.
-                  (fn []
-                    (Thread/sleep 1000)
-                    (let [server (.bind context nil port)
-                          iter (.recv server 0 0)
-                          resp (.next iter)]
-                      (is (= task (.task resp)))
-                      (is (= req_msg (String. (.message resp))))
-                      (.close server))))]
-      (.start server)
-      (wait-until-ready [server client])
-      (.send client task (.getBytes req_msg))
+        server (Thread.
+                 (fn []
+                   (Thread/sleep 100)
+                   (let [server (.bind context nil port)]
+                     (register-callback (fn [message] (reset! resp message)) server))))]
+    (.start server)
+    (wait-until-ready [server client])
+    (.send client task (.getBytes req_msg))
 
-      (.join server)
-      (.close client)
-      (.term context)))
+    (wait-for-not-nil resp)
+    (is (= task (.task @resp)))
+    (is (= req_msg (String. (.message @resp))))
+
+    (.join server)
+    (.close client)
+    (.term context)))
 
 (deftest test-server-delayed
  (let [storm-conf {STORM-MESSAGING-TRANSPORT "backtype.storm.messaging.netty.Context"
@@ -211,28 +231,24 @@
   (log-message "5. test batch")
   (let [num-messages 100000
         _ (log-message "Should send and receive many messages (testing with " num-messages " messages)")
+        resp (ArrayList.)
+        received (atom 0)
         context (TransportFactory/makeContext storm-conf)
         port (available-port 6700)
         server (.bind context nil port)
+        _ (register-callback (fn [message] (.add resp message) (swap! received inc)) server)
         client (.connect context nil "localhost" port)
         _ (wait-until-ready [server client])]
-    (doseq [num  (range 1 num-messages)]
+    (doseq [num (range 1 num-messages)]
       (let [req_msg (str num)]
         (.send client task (.getBytes req_msg))))
 
-    (let [resp (ArrayList.)
-          received (atom 0)]
-      (while (< @received (- num-messages 1))
-        (let [iter (.recv server 0 0)]
-          (while (.hasNext iter)
-            (let [msg (.next iter)]
-              (.add resp msg)
-              (swap! received inc)
-              ))))
-      (doseq [num  (range 1 num-messages)]
+    (while-timeout TEST-TIMEOUT-MS (< (.size resp) (- num-messages 1)) (log-message (.size resp) " " num-messages) (Thread/sleep 10))
+
+    (doseq [num  (range 1 num-messages)]
       (let [req_msg (str num)
             resp_msg (String. (.message (.get resp (- num 1))))]
-        (is (= req_msg resp_msg)))))
+        (is (= req_msg resp_msg))))
 
     (.close client)
     (.close server)
@@ -274,20 +290,18 @@
                       TOPOLOGY-TUPLE-SERIALIZER "backtype.storm.serialization.types.ListDelegateSerializer"
                       TOPOLOGY-FALL-BACK-ON-JAVA-SERIALIZATION false
                       TOPOLOGY-SKIP-MISSING-KRYO-REGISTRATIONS false}
+          resp (atom nil)
           context (TransportFactory/makeContext storm-conf)
           port (available-port 6700)
           client (.connect context nil "localhost" port)
           _ (.send client task (.getBytes req_msg))
           server (.bind context nil port)
+          _ (register-callback (fn [message] (reset! resp message)) server)
           _ (wait-until-ready [server client])
-          _ (.send client task (.getBytes req_msg))
-          iter (future (.recv server 0 0))
-          resp (deref iter 5000 nil)
-          resp-val (if resp (.next resp) nil)]
-      (is resp-val)
-      (when resp-val
-        (is (= task (.task resp-val)))
-        (is (= req_msg (String. (.message resp-val)))))
+          _ (.send client task (.getBytes req_msg))]
+      (wait-for-not-nil resp)
+      (is (= task (.task @resp)))
+      (is (= req_msg (String. (.message @resp))))
       (.close client)
       (.close server)
       (.term context)))